Added .jekyll-cache to .gitignore
While implementing the docker support for development environment
setup
(https://github.com/InnerSourceCommons/innersourcecommons.org/issues/91)
I noticed that the `.jekyll-cache` folder was created, and this folder
was not on `.gitignore`.
I could not find documentation about it, but some references that it
should be excluded form git:
-
https://github.com/github/gitignore/commit/ea2aa40cb4620e334a87e8494e03d0e7d1ec11df#diff-3b5726570b981bf3e11c93a6e7bc3ff9
- https://github.com/github/gitignore/pull/2540

It does not seem like content that belongs to the repo, so I
added to the ignore list.
I wonder if other people also have this folder being generated, or
if it is a side-effect of the Docker setup somehow.